make exposuredimension universal, add exposure to con_context
